refactor: Rename targets-xxx to target-xxx
