+++ pkg/webui/ui/src/components/Router.tsx
@@ -3,7 +3,7 @@ import App from "./App";
 import { ErrorMessageCard } from "./ErrorMessage";
 import { Box } from "@mui/material";
 import React, { useEffect } from "react";
-import { TargetsView } from "./targets-view/TargetsView";
+import { TargetsView } from "./target-view/TargetsView";
 
 function ErrorPage() {
     const error = useRouteError() as any;
