+++ pkg/webui/ui/src/components/target-list-view/TargetsListView.tsx
@@ -7,14 +7,14 @@ import { DataGrid, GridColDef, gridStringOrNumberComparator } from '@mui/x-data-
 import { getLastPathElement } from "../../utils/misc";
 import Tooltip from "@mui/material/Tooltip";
 import { Approval, Toys } from "@mui/icons-material";
-import { ReconcilingIcon } from "../targets-view/ReconcilingIcon";
-import { StatusIcon } from "../targets-view/StatusIcon";
-import { TargetActionMenu } from "../targets-view/TargetActionMenu";
+import { ReconcilingIcon } from "../target-view/ReconcilingIcon";
+import { StatusIcon } from "../target-view/StatusIcon";
+import { TargetActionMenu } from "../target-view/TargetActionMenu";
 import { CommandResultStatusLine } from "../command-result/CommandResultStatusLine";
 import { Since } from "../Since";
 import { ScrollingTextLine } from "../ScrollingTextLine";
-import { ClusterIcon } from "../targets-view/ClusterIcon";
-import { CommandTypeIcon } from "../targets-view/CommandTypeIcon";
+import { ClusterIcon } from "../target-view/ClusterIcon";
+import { CommandTypeIcon } from "../target-view/CommandTypeIcon";
 import Divider from "@mui/material/Divider";
 import { TargetCard } from "../target-cards-view/TargetCard";
 import { CommandResultCard } from "../command-result/CommandResultCard";
