+++ pkg/webui/ui/src/components/target-view/ClusterIcon.tsx
