+++ pkg/webui/ui/src/components/target-view/CommandTypeIcon.tsx
