+++ pkg/webui/ui/src/components/target-view/DiscriminatorIcon.tsx
