+++ pkg/webui/ui/src/components/target-view/TargetTypeIcon.tsx
